+++ ltexls-core/src/test/java/org/bsplines/ltexls/parsing/markdown/MarkdownAnnotatedTextBuilderTest.java
@@ -52,6 +52,16 @@ public class MarkdownAnnotatedTextBuilderTest {
     assertPlainText(
         "This is a test: $`E = mc^2`$.\n\n```math\na^2 + b^2 = c^2\n```\n\nThis is another test.\n",
         "This is a test: Dummy0.\n\n\n\n\n\nThis is another test.\n");
+    assertPlainText(
+        "This is a test.\n"
+        + "\n"
+        + "| First Column | Second Column |\n"
+        + "| ------------ | ------------- |\n"
+        + "| Interesting  | Super         |\n"
+        + "\n"
+        + "This is another sentence.\n",
+        "This is a test.\n\nFirst Column Second Column\n\nInteresting Super\n\n"
+        + "This is another sentence.\n");
 
     Map<String, String> markdownNodes = new HashMap<>();
     markdownNodes.put("Code", "default");